requirements.txt
streamlit
opencv-python
numpy
torch
pillow